Final styling

+++ client/src/pages/HistoryPage.js
@@ -3,6 +3,7 @@ import Sidebar from './components/Sidebar';
 import { _, Grid } from 'gridjs-react';
 import * as FaIcons from 'react-icons/fa';
 import * as MdIcons from 'react-icons/md';
+import Swal from 'sweetalert2';
 import jsPDF from 'jspdf';
 import 'jspdf-autotable';
 import { newMonth, numberMonth, month } from '../helpers/helpers';
@@ -68,6 +69,26 @@ function HistoryPage() {
     };
     dispatch(createExpenses(newDataExpense));
     handleCloseAddForm();
+    const Toast = Swal.mixin({
+      toast: true,
+      position: 'top-end',
+      showConfirmButton: false,
+      timer: 2000,
+      timerProgressBar: true,
+      didOpen: (toast) => {
+        toast.addEventListener('mouseenter', Swal.stopTimer);
+        toast.addEventListener('mouseleave', Swal.resumeTimer);
+      },
+    });
+
+    Toast.fire({
+      icon: 'success',
+      title: 'Add is Successfully',
+    });
+    setExpenseAddTitle('');
+    setExpenseAddMonth('');
+    setExpenseAddYear('');
+    setExpenseAddTotal('');
   };
 
   const updateExpenseTransaction = () => {
@@ -79,10 +100,53 @@ function HistoryPage() {
     };
     dispatch(updateExpenses(expenseUpdateId, updateDataExpense));
     handleCloseUpdateForm();
+    const Toast = Swal.mixin({
+      toast: true,
+      position: 'top-end',
+      showConfirmButton: false,
+      timer: 2000,
+      timerProgressBar: true,
+      didOpen: (toast) => {
+        toast.addEventListener('mouseenter', Swal.stopTimer);
+        toast.addEventListener('mouseleave', Swal.resumeTimer);
+      },
+    });
+
+    Toast.fire({
+      icon: 'success',
+      title: 'Updeted Successfully',
+    });
   };
 
   const handelDeleteExpense = (id) => {
-    dispatch(deleteExpense(id));
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(deleteExpense(id));
+        const Toast = Swal.mixin({
+          toast: true,
+          timer: 2000,
+          timerProgressBar: true,
+          showConfirmButton: false,
+          position: 'top-end',
+          didOpen: (toast) => {
+            toast.addEventListener('mouseenter', Swal.stopTimer);
+            toast.addEventListener('mouseleave', Swal.resumeTimer);
+          },
+        });
+        Toast.fire({
+          icon: 'success',
+          title: 'Expense sucessfully deleted.',
+        });
+      }
+    });
   };
 
   let newDataExpenseBar = [];
@@ -175,15 +239,16 @@ function HistoryPage() {
                   padding: '20px',
                 }}
               >
-                <h3
+                <h2
                   className='text-center mb-3'
                   style={{
                     padding: '10px',
                     fontWeight: 'bold',
+                    color: '#343F56',
                   }}
                 >
                   Expense Table
-                </h3>
+                </h2>
 
                 <div style={{ alignSelf: 'flex-center' }}>
                   <Button
@@ -465,7 +530,7 @@ function HistoryPage() {
                   padding: '20px',
                 }}
               >
-                <h3
+                <h2
                   className='text-center mb-3'
                   style={{
                     // border: 'solid',
@@ -475,7 +540,7 @@ function HistoryPage() {
                   }}
                 >
                   Income Table
-                </h3>
+                </h2>
 
                 <div style={{ alignSelf: 'flex-center' }}>
                   <Button
